requirements.txt
gradio>=4.0.0
transformers>=4.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
accelerate>=0.20.0
tokenizers>=0.13.0
safetensors>=0.3.0
python-dotenv>=1.0.0
psutil>=5.9.0 